+++ common/src/main/java/com/avp/common/gameplay/ai/goal/combat/DelayedAttackGoal.java
@@ -1,23 +1,22 @@
 package com.avp.common.gameplay.ai.goal.combat;
 
-import com.alien.common.gameplay.entity.living.alien.Alien;
-import com.alien.common.util.AlienPredicates;
+import com.lib.common.data.Cooldown;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.PathfinderMob;
 import net.minecraft.world.entity.ai.goal.MeleeAttackGoal;
+import org.jetbrains.annotations.NotNull;
 
+import java.time.Duration;
 import java.util.EnumSet;
 
 public class DelayedAttackGoal extends MeleeAttackGoal {
 
-    private final int delayTicksBeforeAttack;
-
     private final Runnable attackAnimationRunnable;
 
-    private int delayBeforeAttack;
+    private final Cooldown attackAnimationCooldown;
 
-    private boolean triggeredAttackAnimation;
+    private boolean ranAttackAnimation;
 
     public DelayedAttackGoal(
         PathfinderMob mob,
@@ -27,44 +26,44 @@ public class DelayedAttackGoal extends MeleeAttackGoal {
         Runnable attackAnimationRunnable
     ) {
         super(mob, speedModifier, bl);
-        this.delayTicksBeforeAttack = delayTicksBeforeAttack;
+        this.attackAnimationCooldown = Cooldown.withCooldownTime("attackAnimationCooldownInTicks", Duration.ofMillis(delayTicksBeforeAttack * 50L));
         this.attackAnimationRunnable = attackAnimationRunnable;
 
         setFlags(EnumSet.of(Flag.MOVE, Flag.LOOK));
     }
 
     @Override
-    public void start() {
-        super.start();
-        this.delayBeforeAttack = 0;
-        this.triggeredAttackAnimation = false;
+    public void tick() {
+        super.tick();
+        attackAnimationCooldown.tick();
+
+        if (
+            // If target is not null
+            mob.getTarget() != null
+                // AND we ran the attack animation.
+                && ranAttackAnimation
+                // AND the animation cooldown has finished
+                && !attackAnimationCooldown.isActive()
+                // AND target is still within melee range
+                && mob.isWithinMeleeAttackRange(mob.getTarget())
+                // AND we still have line of sight of the target
+                && mob.getSensing().hasLineOfSight(mob.getTarget())
+        ) {
+            resetAttackCooldown();
+            mob.swing(InteractionHand.MAIN_HAND);
+            mob.doHurtTarget(mob.getTarget());
+            this.ranAttackAnimation = false;
+        }
     }
 
     @Override
-    protected void checkAndPerformAttack(LivingEntity target) {
-        if (canPerformAttack(target)) {
-            if (delayBeforeAttack > 0) {
-                delayBeforeAttack--;
-
-                if (delayBeforeAttack == delayTicksBeforeAttack && !triggeredAttackAnimation) {
-                    attackAnimationRunnable.run();
-                    this.triggeredAttackAnimation = true;
-                }
-            } else {
-                resetAttackCooldown();
-                mob.swing(InteractionHand.MAIN_HAND);
-
-                if (mob instanceof Alien alien) {
-                    var detectionRange = 5.0;
-                    AlienPredicates.prioritizeAndAttack(alien, detectionRange);
-                } else {
-                    mob.doHurtTarget(target);
-                }
-                this.triggeredAttackAnimation = false;
-            }
-        } else {
-            this.delayBeforeAttack = adjustedTickDelay(10);
-            this.triggeredAttackAnimation = false;
+    protected void checkAndPerformAttack(@NotNull LivingEntity target) {
+        if (!ranAttackAnimation && canPerformAttack(target)) {
+            // Play the animation.
+            attackAnimationRunnable.run();
+            this.ranAttackAnimation = true;
+            // Reset the cooldown.
+            attackAnimationCooldown.reset();
         }
     }
 }