+++ CHANGELOG.md
@@ -84,6 +84,7 @@
 - Fixed inconsistent potion effect behavior with embryos born from hosts eating chorus fruit.
 - Fixed aberrant chitin armors not counting towards chitin armor advancements.
 - Fixed aliens spreading resin while underwater.
+- Fixed alien attack animations not playing properly.
 
 ## 🛠 Data Pack
 - Added `#avp:runner_hosts` entity tag.